requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.7
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.15.1
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
